+++ src/main.ts
@@ -6,7 +6,7 @@ import setupIpcComs from './utils/ipcComms';
 import { buildContextMenu, buildMenuBar } from './utils/menuUtil';
 import initSentry from './utils/sentry';
 import { isDev } from './utils/common';
-import { initWatcher } from './utils/watch';
+import { initWatcher } from './services/chokidar';
 
 if (isDev) {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
